backend/pyproject.toml

classifiers=[
    "License :: OSI Approved :: MIT License",
    "Topic :: Scientific/Engineering",
    "Development Status :: 4 - Beta",
]


[tool.poetry]
name = "multimodalexplorer"
version = "0.1.0"
description = "an embedding space explorer"
authors = ["Facebook AI Research"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
fastapi = "^0.110.0"
uvicorn = {extras = ["standard"], version = "^0.27.1"}
pydantic = "^2.6.3"
datasets = "^2.18.0"
umap-learn = "^0.5.6"
pandas = "2.0.3"
faiss-cpu = "^1.8.0"
argparse = "^1.4.0"
sonar-space = "0.2.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
